Created GameView, Board Class, and Cell Classes. I also verified that the GameView class works (It uses the other ones so I think they work too).
Had to make the Cell enum house integers as opposed to booleans because it would, for some reason, assign all of the Cells to be the top cell which was DiskP1 so I had to make them all random integers.

I put in a dummy flip class in the Cell. I honestly couldn't figure out what that was supposed to be since I would have made it do the same thing as the "fill" class.