I'm not done yet. Put this here for padraic.
I made it so the pick game page controller and view work. I had to work around a bunch of problems to make it work.

The weird import statement is because the controller and view both need to communicate with each other. I know this is a little weird but everywhere I looked online said that with GUI's the view notifies the controller of changes. Because of this, the view has to communicate with the controller and the controller has to tell the view what to do. Took me a long time to figure out the fix for this weird import "error"